% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fucntions.R
\name{cal.index}
\alias{cal.index}
\title{Title index2 Calculating performances indexes}
\usage{
cal.index(data, var, alpha)
}
\arguments{
\item{data}{A data.frame has at least two columns.
(gene name and p-values (or adjusted p-values))}

\item{var}{The colname of p-values or adjusted p-values.}

\item{alpha}{Threshold.}
}
\value{
A numeric vector.
}
\description{
Title index2 Calculating performances indexes
}
