% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fucntions.R
\name{filter.gene}
\alias{filter.gene}
\title{Title   filter.gene
This function filters the non-expressed genes and low-proportion expressed genes.}
\usage{
filter.gene(object, thr)
}
\arguments{
\item{object}{: A seurat object.}

\item{thr}{A number.
If thr is between 0 and 1, it indicates the least ratio of each gene expressed in the spots.
If thr is greater than 1, it indicates the minimum number of each gene expressed in the spots.}
}
\value{
A seurat object.
}
\description{
Title   filter.gene
This function filters the non-expressed genes and low-proportion expressed genes.
}
