% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gradient.zi}
\alias{gradient.zi}
\title{Title Simulation data. Genes with gradient expression pattern.}
\usage{
gradient.zi(
  lambda = 0.5,
  spots,
  se,
  ns,
  type = "ZINB",
  ns.p,
  ns.size = 0.2,
  ns.mu,
  a = 1.5,
  b = 0,
  grid = 10,
  grid.part = 0.5,
  outlier = FALSE
)
}
\arguments{
\item{lambda}{The spots detection efficiency.It should be between 0 and 1.
The default value is 0.5.}

\item{spots}{The number of spots.}

\item{se}{The number of  spatially variable genes (SVGs).}

\item{ns}{The number of  non-SVGs.}

\item{type}{A character.'ZINB' or 'ZIP. The default value is ZINB.}

\item{ns.p}{A number. The zero proportion of zero generation process of
non-SVGs and SVGs in the non-streak area.}

\item{ns.size}{For non-SVGs and SVGs in the non-streak area,
the size para in the NB distribution.}

\item{ns.mu}{For non-SVGs and SVGs in the non-streak area, the lambda para in the poisson distribution}

\item{a}{The strength parameter of gradient pattern.}

\item{b}{The strength parameter of gradient pattern.}

\item{grid}{The strength parameter of gradient pattern.}

\item{grid.part}{The proportion of gradient area.}

\item{outlier}{A numeric value should be between 0 and 1.}

\item{prop}{The proportion of streak area.}
}
\value{
A data.frame of gene expression counts and coordinates.
}
\description{
Title Simulation data. Genes with gradient expression pattern.
}
