% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0907_new_simulation.R
\name{irreg.zi}
\alias{irreg.zi}
\title{Title irreg.zi
new simulation, patterns from irregular png. spots from the Poisson process.}
\usage{
irreg.zi(
  lambda = 0.7,
  spots,
  se,
  ns,
  type,
  se.p,
  se.size = 0.2,
  se.mu,
  ns.p,
  ns.size = 0.2,
  ns.mu,
  ptn,
  png_dir,
  outlier = FALSE
)
}
\arguments{
\item{lambda}{The spots detection efficiency.It should be between 0 and 1.
The default value is 0.5.}

\item{spots}{The number of spots.}

\item{se}{The number of  spatially variable genes (SVGs).}

\item{ns}{The number of  non-SVGs.}

\item{type}{A character.'ZINB' or 'ZIP. The default value is ZINB.}

\item{se.p}{A number. The zero proportion of zero generation process of
SVGs in the streak area.}

\item{se.size}{The size para of SVGs in the NB distribution.}

\item{se.mu}{For SVGs, the lambda para in the poisson distribution
or the mu para in the NB distribution.}

\item{ns.p}{A number. The zero proportion of zero generation process of
non-SVGs and SVGs in the non-streak area.}

\item{ns.size}{For non-SVGs and SVGs in the non-streak area,
the size para in the NB distribution.}

\item{ns.mu}{For non-SVGs and SVGs in the non-streak area, the lambda para in the poisson distribution
or the mu para in the NB distribution.}

\item{ptn}{The file name of the pattern png.}

\item{png_dir}{The dir of the png files.}

\item{outlier}{A numeric value should be between 0 and 1.}

\item{prop}{The proportion of streak area.}
}
\value{
A data.frame of gene expression counts and coordinates.
}
\description{
Title irreg.zi
new simulation, patterns from irregular png. spots from the Poisson process.
}
