% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0907_new_simulation.R
\name{new.norm}
\alias{new.norm}
\title{Title new.norm}
\usage{
new.norm(
  lambda = 0.7,
  spots,
  se,
  ns,
  se.p,
  ns.p,
  se.sd = 0.2,
  se.mu,
  ns.sd = 0.2,
  ns.mu,
  ptn,
  png_dir,
  outlier = FALSE
)
}
\arguments{
\item{lambda}{The spots detection efficiency.It should be between 0 and 1.
The default value is 0.5.}

\item{spots}{The number of spots.}

\item{se}{The number of  spatially variable genes (SVGs).}

\item{ns}{The number of  non-SVGs.}

\item{se.p}{A number. The zero proportion of zero generation process of
SVGs in the streak area.}

\item{ns.p}{A number. The zero proportion of zero generation process of
non-SVGs and SVGs in the non-streak area.}

\item{se.sd}{the sd of the SVG.}

\item{se.mu}{the mean of the SVG.}

\item{ns.sd}{the sd of the non-SVG.}

\item{ns.mu}{the mean of the non-SVG.}

\item{ptn}{The file name of the pattern png.}

\item{png_dir}{The dir of the png files.}

\item{outlier}{A numeric value should be between 0 and 1.}

\item{prop}{The proportion of streak area.}
}
\value{
A data.frame of gene expression counts and coordinates.
}
\description{
pattern from png, zero-inflater+normal distribution, 
 spots from the Poisson process.
}
