% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fucntions.R
\name{scale.count}
\alias{scale.count}
\title{Title:scale.count  (original name: simu.count)
Transformation of expression counts.
To adjust the effects from library size between sample spots and outliers of gene expression.}
\usage{
\method{scale}{count}(data, qh = 0.985)
}
\arguments{
\item{data}{A data.frame or matrix of expression counts (spots x genes)}

\item{qh}{A numeric value should be between 0 and 1.}
}
\value{
A data.frame or matrix of adjusted expression counts (spots x genes)
}
\description{
Title:scale.count  (original name: simu.count)
Transformation of expression counts.
To adjust the effects from library size between sample spots and outliers of gene expression.
}
