% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fucntions.R
\name{simu.scale}
\alias{simu.scale}
\title{Title:basic functions}
\usage{
simu.scale(data)
}
\arguments{
\item{data}{A data.frame or matrix of expression counts}
}
\value{
A data.frame or matrix.
}
\description{
Title:basic functions
}
