% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0907_new_basic.R
\name{to_dt}
\alias{to_dt}
\title{Title to_da}
\usage{
to_dt(matrix_data)
}
\arguments{
\item{matrix_data}{p*n matrix. From a .png file.}
}
\value{
A data frame with 3 columns (row,col,exp).
}
\description{
Transformation of p*n matrix to a data frame with 3 columns (row,col,exp).
}
