% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0907_with_scGCO.R
\name{toGCO_2}
\alias{toGCO_2}
\title{Title toGCO_2}
\usage{
toGCO_2(data)
}
\arguments{
\item{data}{input, a data.frame, top 2 columns are row and col.}
}
\value{
a .tsv file for scGCO.
}
\description{
Generating tsv for scGCO.
}
