% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_functions.R
\name{ts.svg2}
\alias{ts.svg2}
\title{Title  ts.svg2  Finding SVG (parallel).}
\usage{
ts.svg2(data, scale = T, noise = T, padj.m = "holm")
}
\arguments{
\item{data}{A data.frame with counts and coordinates.(spots x genes)
The first two columns are coordinates. The first two colnames must be 'row' and 'col'.}

\item{scale}{A logic value.T/F
If scale==T, the expression counts matrix runs function scale.count.}

\item{noise}{A logic value.T/F.
If noise==F, the SVG-rank considers only the p-value and adjusted p-value.
If noise==T, the SVG-rank takes into account the p-value,adjusted p-valus
 and the distribution characteristics of the genes.}

\item{padj.m}{A character string.
The adjusted method of p-values. The default value is 'holm'.}
}
\value{
A data.frame. All genes' p values, adjusted p values and SVG-rank.
}
\description{
Title  ts.svg2  Finding SVG (parallel).
}
